+++ db/access.php
@@ -90,6 +90,18 @@ $capabilities = [
         ],
     ],
 
+    'mod/vimipad:peerreview' => [
+        'riskbitmask' => RISK_SPAM,
+        'captype' => 'write',
+        'contextlevel' => CONTEXT_MODULE,
+        'archetypes' => [
+            'student' => CAP_ALLOW,
+            'teacher' => CAP_ALLOW,
+            'editingteacher' => CAP_ALLOW,
+            'manager' => CAP_ALLOW,
+        ],
+    ],
+
     'mod/vimipad:grade' => [
         'captype' => 'write',
         'contextlevel' => CONTEXT_MODULE,